split Goseconfiguration out to new file
