+++ tests/setup_teardown.rs
@@ -6,8 +6,8 @@ use serial_test::serial;
 
 mod common;
 
+use goose::config::GooseConfiguration;
 use goose::prelude::*;
-use goose::GooseConfiguration;
 
 // Paths used in load tests performed during these tests.
 const INDEX_PATH: &str = "/";
